# backend/peerprep-question-service/src/main/java/com/peerprep/peerprepquestionservice/controller/QuestionController.java
package com.peerprep.peerprepquestionservice.controller;

import com.peerprep.peerprepcommon.dto.question.*;
import com.peerprep.peerprepquestionservice.exception.NoSuitableQuestionException;
import com.peerprep.peerprepquestionservice.exception.QuestionNotFoundException;
import com.peerprep.peerprepquestionservice.service.QuestionService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@CrossOrigin(origins = {"http://localhost:8080/", "https://api-gateway-dot-peerprep-399116.as.r.appspot.com/"})
@RestController
@RequiredArgsConstructor
public class QuestionController {


    private final QuestionService questionService;

    @PostMapping
    public ResponseEntity<String> createQuestion(@RequestBody @Valid final CreateQuestionRequest request) {
        String id = questionService.createQuestion(request);
        return ResponseEntity.status(201).body(id);
    }

    @GetMapping
    public ResponseEntity<List<QuestionOverview>> getQuestions() {
        List<QuestionOverview> questions = questionService.getQuestions();
        return ResponseEntity.ok(questions);
    }

    @GetMapping("/{id}")
    public ResponseEntity<QuestionResponse> getQuestion(@PathVariable final String id) {
        QuestionResponse question = questionService.getQuestion(id);
        return ResponseEntity.ok(question);
    }

    @PostMapping("/random")
    public ResponseEntity<String> getRandomQuestion(@RequestBody RandomQuestionRequest request) {
        try {
            String id = questionService.getRandomQuestion(request);
            return ResponseEntity.ok(id);
        } catch (NoSuitableQuestionException e) {
            return ResponseEntity.ok("-1");
        }
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteQuestion(@PathVariable final String id) {
        questionService.deleteQuestion(id);
        return ResponseEntity.ok().build();
    }

    @PatchMapping("/{id}")
    public ResponseEntity<Void> updateQuestion(@PathVariable final String id, @RequestBody @Valid final UpdateQuestionRequest request) {
        questionService.updateQuestion(id, request);
        return ResponseEntity.ok().build();
    }

    @ExceptionHandler(QuestionNotFoundException.class)
    public ResponseEntity<String> handleQuestionNotFoundException(QuestionNotFoundException e) {
        String msg = String.format("Question %s not found", e.getId());
        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(msg);
    }
}